requirements.txt
pandas
scikit-learn
mlflow
joblib
pyyaml
kaggle
dvc
